refactor: domain interface
