+++ src/main/java/com/fullcar/carpool/application/form/FormService.java
@@ -1,10 +1,10 @@
 package com.fullcar.carpool.application.form;
 
-import com.fullcar.carpool.application.Alarm.AlarmService;
 import com.fullcar.carpool.domain.carpool.Carpool;
 import com.fullcar.carpool.domain.carpool.CarpoolId;
 import com.fullcar.carpool.domain.carpool.CarpoolRepository;
 import com.fullcar.carpool.domain.form.*;
+import com.fullcar.carpool.infra.NotificationClient;
 import com.fullcar.carpool.presentation.form.dto.request.FormRequestDto;
 import com.fullcar.carpool.presentation.form.dto.request.FormUpdateDto;
 import com.fullcar.carpool.presentation.form.dto.response.FormResponseDto;
@@ -29,7 +29,7 @@ public class FormService {
     private  final CarpoolRepository carpoolRepository;
     private final MemberRepository memberRepository; //TODO: Event 기반으로 변경 필요.
     private final FormMapper formMapper;
-    private final AlarmService alarmService;
+    private final NotificationClient pushNotificationClient;
 
     @Transactional
     public FormResponseDto requestForm(Member member, CarpoolId carpoolId, FormRequestDto formRequestDto) {
@@ -53,7 +53,7 @@ public class FormService {
         }
 
         Form form = formMapper.toEntity(member, carpoolId, formRequestDto);
-        alarmService.sendNotification(driver.getNickname(), driver.getDeviceToken(), "탑승 요청이 들어왔어요!", "탑승자 정보를 확인하고 승인해 주세요🚘");
+        pushNotificationClient.sendNotification(driver.getNickname(), driver.getDeviceToken(), "탑승 요청이 들어왔어요!", "탑승자 정보를 확인하고 승인해 주세요🚘");
 
         return formMapper.toDto(
                 formRepository.saveAndFlush(form),
@@ -109,12 +109,12 @@ public class FormService {
         if (formUpdateDto.getFormState() == FormState.ACCEPT) {
             String title = "카풀 매칭에 성공했어요!";
             String body = "운전자 정보를 확인해 주세요🚘";
-            alarmService.sendNotification(passenger.getNickname(), passenger.getDeviceToken(), title, body);
+            pushNotificationClient.sendNotification(passenger.getNickname(), passenger.getDeviceToken(), title, body);
         }
         else if (formUpdateDto.getFormState() == FormState.REJECT) {
             String title = "카풀 매칭에 실패했어요.";
             String body = "다른 카풀을 찾아볼까요?🥹";
-            alarmService.sendNotification(passenger.getNickname(), passenger.getDeviceToken(), title, body);
+            pushNotificationClient.sendNotification(passenger.getNickname(), passenger.getDeviceToken(), title, body);
         }
 
         return formMapper.toDetailDto(
